linux-stable: add defconfig for cogip-raspberrypi4-64
This defconfig is optimized for Raspberry Pi4 Model B, building
a 64 bits kernel image with VC4 GPU support.
Almost no driver is built as module except brcmfmac WiFi driver
for firmware loading purpose.
Autofs4 is also built as a module because systemd is trying to load
it, leading to errors if built-in.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>